Window info issues workaround
When using Selenium2Library with Firefox 24, it's possible that
WebDriver returns a WebElement instance as current window id. This
causes TypeError, because len() could not be applied to non-iterable
object.

In order to fix this, we have to check first, whether the returned value
is instance of WebElement or not.